requirement.txt
django-mptt >= 0.11.0
django-tagging >= 0.5.0
beautifulsoup4 >= 4.8.2
mots-vides >= 2015.5.11
pyparsing >= 2.4.6
regex >= 2020.2.20
django-contrib-comments >= 1.9.2
django-xmlrpc >= 0.1.8
